Appointment issues

+++ src/Screens/Components/CommonProfileSec/DeleteAccount.js
@@ -10,7 +10,7 @@ import axios from "axios";
 import Loader from "Screens/Components/Loader/index";
 import { confirmAlert } from "react-confirm-alert"; // Import
 import "react-confirm-alert/src/react-confirm-alert.css"; // Import css
-import { getLanguage} from "translations/index";
+import { getLanguage } from "translations/index";
 import { commonHeader, commonCometDelHeader } from 'component/CommonHeader/index';
 
 class Index extends Component {
@@ -26,34 +26,20 @@ class Index extends Component {
   }
 
   //FOR DELETING THE USER
-  DeleteUser=(deletekey, profile_id, bucket)=>{
+  DeleteUser = (deletekey, profile_id, bucket) => {
+    var UserId = deletekey;
+    let translate = getLanguage(this.props.stateLanguageType)
+    let { ok } = translate;
     this.setState({ loaderImage: true });
     const user_token = this.props.stateLoginValueAim.token;
-    axios.delete(sitedata.data.path + '/admin/deleteUser/' + deletekey+'?bucket='+bucket, commonHeader(user_token))
-        .then((response) => {
-            this.setState({ loaderImage: false });
-            var data = JSON.stringify({"permanent":true});
-            var config = {
-              method: 'delete',
-              url: 'https://api-eu.cometchat.io/v2.0/users/'+profile_id.toLowerCase(),
-              headers: commonCometDelHeader(),
-              data : data
-            };
-            
-            axios(config)
-            .then(function (response) { })
-            .catch(function (error) { });
-            this.props.history.push('/');
-
-        }).catch((error) => {});
-}
-
-  //Dialog for deleting accoount 
-  DeleteAccount=()=>{
-      if(!this.state.notmatch){
-        let translate = getLanguage(this.props.stateLanguageType)
-        let { cancel, ok, reallyWantDelete, looseAllData } = translate;
-        confirmAlert({
+    // axios.delete(sitedata.data.path + '/admin/deleteUser/' + deletekey+'?bucket='+bucket, commonHeader(user_token))
+    axios.delete(sitedata.data.path + '/admin/deleteUser/' + UserId, commonHeader(user_token))
+      .then((response) => {
+        if (response &&
+          response.data &&
+          response.data.messages === "video chat account exists, deactivate that first"
+        ) {
+          confirmAlert({
             customUI: ({ onClose }) => {
               return (
                 <div
@@ -65,30 +51,82 @@ class Index extends Component {
                       : "react-confirm-alert-body"
                   }
                 >
-                  <h1 className="deletewarninghead">{reallyWantDelete}</h1>
-                  <p className="deletewarning">{looseAllData}</p>
-                  <div className="react-confirm-alert-button-group">
+                  <p className="deletewarning">You are registered as a video conference account, for delete an account from Aimedis first you need to deactivate that account, After that you can delete this Aimedis account !</p>
+                  <div className="react-confirm-alert-button-group react-confirm-alert-button-group1">
                     <button
                       onClick={() => {
-                        this.DeleteUser(this.state.Current_state._id, this.state.Current_state.profile_id, this.state.Current_state.bucket);
                         onClose();
                       }}
                     >
                       {ok}
-                    </button>
-                    <button
-                      onClick={() => {
-                        onClose();
-                      }}
-                    >
-                      {cancel}
                     </button>
                   </div>
                 </div>
               );
             },
           });
-      }
+          this.setState({ loaderImage: false });
+        }
+
+        else {
+          this.setState({ loaderImage: false });
+          var data = JSON.stringify({ "permanent": true });
+          var config = {
+            method: 'delete',
+            url: 'https://api-eu.cometchat.io/v2.0/users/' + profile_id.toLowerCase(),
+            headers: commonCometDelHeader(),
+            data: data
+          };
+
+          axios(config)
+            .then(function (response) { })
+            .catch(function (error) { });
+          this.props.history.push('/');
+        }
+      }).catch((error) => { });
+  }
+
+  //Dialog for deleting accoount 
+  DeleteAccount = () => {
+    if (!this.state.notmatch) {
+      let translate = getLanguage(this.props.stateLanguageType)
+      let { cancel, ok, reallyWantDelete, looseAllData } = translate;
+      confirmAlert({
+        customUI: ({ onClose }) => {
+          return (
+            <div
+              className={
+                this.props.settings &&
+                  this.props.settings.setting &&
+                  this.props.settings.setting.mode === "dark"
+                  ? "dark-confirm react-confirm-alert-body"
+                  : "react-confirm-alert-body"
+              }
+            >
+              <h1 className="deletewarninghead">{reallyWantDelete}</h1>
+              <p className="deletewarning">{looseAllData}</p>
+              <div className="react-confirm-alert-button-group">
+                <button
+                  onClick={() => {
+                    this.DeleteUser(this.state.Current_state._id, this.state.Current_state.profile_id, this.state.Current_state.bucket);
+                    onClose();
+                  }}
+                >
+                  {ok}
+                </button>
+                <button
+                  onClick={() => {
+                    onClose();
+                  }}
+                >
+                  {cancel}
+                </button>
+              </div>
+            </div>
+          );
+        },
+      });
+    }
   }
   //For Change Password State For version V4
   ChangePass = (e) => {
@@ -118,7 +156,7 @@ class Index extends Component {
     this.setState({ Password: state });
   };
 
- 
+
   render() {
     let translate = getLanguage(this.props.stateLanguageType)
     let {
@@ -136,39 +174,39 @@ class Index extends Component {
         {this.state.notmatch && (
           <div className="err_message">{current_pass_not_match}</div>
         )}
-      
+
         <Grid container direction="row" alignItems="center" spacing={2}>
           <Grid item xs={12} md={5}>
-              
+
             <Grid className="chngPasswrd">
               <h2>{delete_your_account}</h2>
               {/* <p>{supportive_text}</p> */}
             </Grid>
             <Grid className="infoSub">
-                <input
+              <input
                 type="submit"
                 className=""
-                onClick={()=>{this.setState({enterpass: true})}}
+                onClick={() => { this.setState({ enterpass: true }) }}
                 value={delete_account}
-                />
+              />
             </Grid>
-            {this.state.enterpass && 
-            <Grid className="genPass current_pass_sec">
+            {this.state.enterpass &&
+              <Grid className="genPass current_pass_sec">
                 <Grid className="genPassInr">
-                    <label>
+                  <label>
                     {Current} {password}
-                    </label>
-                    <input
-                      type="password"
-                      name="current_pass"
-                      onChange={(e)=>this.ChangePass(e)}
-                    />
-                    <input
-                      type="submit"
-                      onClick={this.DeleteAccount}
-                      value={Submit} 
-                    />
-                </Grid>   
+                  </label>
+                  <input
+                    type="password"
+                    name="current_pass"
+                    onChange={(e) => this.ChangePass(e)}
+                  />
+                  <input
+                    type="submit"
+                    onClick={this.DeleteAccount}
+                    value={Submit}
+                  />
+                </Grid>
               </Grid>}
           </Grid>
           <Grid item xs={12} md={7}></Grid>
